+++ src/clone.rs
@@ -4,7 +4,6 @@ use fehler::throws;
 
 #[throws(eyre::Error)]
 pub(crate) fn clone_repo(repo_url: &str, dst: &PathBuf) {
-    println!("Cloning repository");
     let mut cmd = std::process::Command::new("git");
     cmd.arg("clone");
     cmd.arg(repo_url);