+++ src/main.rs
@@ -8,11 +8,9 @@ mod repo;
 
 use clap::Parser;
 
-use crate::repo::get_repo;
 fn main() -> Result<(), Box<dyn Error>> {
     let args = cli::Args::parse();
-    let repo = get_repo(args.remote.as_str())?;
-    return app::run_app(repo, args.branches, args.recent);
+    return app::run_app(args.remote, args.branches, args.recent);
 }
 
 #[cfg(test)]